Move common utils to common package
